Add extractors for future event response data
This commit includes the addition of new extractors for processing future event response data. These extractors have been designed to handle lower and higher frequency data, as well as information related to the activation of new systems. Changes are also incorporated in the main function to utilize these new extractors.